Menu and loading Playing Modes
A main menu was implemented to select the two player modes and an exit options.

The 2 player mode was implemented by creating 2 circles player. The movement keys are now Horizontal and Vertical, not Jump. The camera follows the player that's ahead killing the one who stays behind if it get out of if. Also the cross obstacle was taken out from the obstacle course, until further improvements

To work on:
- integration of the cross obstacle in the 2 player mode
- a death menu for both modes