Game Over Menu, Camera Move and Barrier
The Game Over Menu arises each time the player dies to an obstacle or barrier. Death by the out of bound condition still needs to be implemented.

Camera Move - implies the fact that the camera starts to move when no player moves further

Barrier is made up of 4 walls of the elements. Each wall must be destroyed by shooting 5 bullets of different element.

Must work on the square duo prefab.